Extract the rate limiting code in a limits package
